requirements.txt
appdirs==1.4.4
astroid==2.4.2
beautifulsoup4==4.9.3
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
click==7.1.2
colorama==0.4.3
cssselect2==0.4.1
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.19
django-bootstrap4==3.0.1
django-filter==2.4.0
django-session-csrf==0.7.1
django-stdimage==5.3.0
django-weasyprint==1.1.0
djangorestframework==3.12.4
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
html5lib==1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nltk==3.6.2
Pillow==8.1.2
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.5.3
Pyphen==0.10.0
python-dotenv==0.14.0
pytz==2021.1
regex==2021.4.4
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
textblob==0.15.3
tinycss2==1.1.0
toml==0.10.1
tqdm==4.61.0
virtualenv==20.0.25
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
